refactor: Split predicate_variations.rs into focused test modules
Refactored the 1,003-line predicate_variations.rs test file into 6
focused modules organized by predicate type, improving maintainability
and code organization.

## Changes

### Module Structure
- Created `predicate_tests/` directory with 6 sub-modules:
  - `in_predicate.rs` (177 lines) - IN/NOT IN tests
  - `like_predicate.rs` (237 lines) - LIKE/NOT LIKE tests
  - `between_predicate.rs` (129 lines) - BETWEEN tests
  - `position_function.rs` (153 lines) - POSITION function tests
  - `trim_function.rs` (220 lines) - TRIM function tests
  - `cast_expression.rs` (107 lines) - CAST expression tests

### Test Preservation
- All 25 original tests preserved and passing
- Zero test functionality changes
- Improved imports using `use super::super::super::*;`

### Documentation
- Added comprehensive `README.md` with:
  - Test coverage matrix for each module
  - SQL:1999 standard compliance notes
  - Edge case documentation
  - Maintenance guidelines
- Enhanced module-level documentation in `mod.rs`

### Benefits
- Each module < 250 lines (meets acceptance criteria)
- Clear separation of concerns by predicate type
- Easier to locate and maintain specific test categories
- Better documentation of SQL standard compliance

Closes #623
